Refactor : 조회 시 검색 결과 없을 시에 예외 처리 추가
