+++ src/test/java/com/runninghi/keyword/command/application/service/CommandKeywordServiceTest.java
@@ -46,7 +46,6 @@ class CommandKeywordServiceTest {
         // when & then
         Assertions.assertThatCode(
                 () -> commandKeywordService.checkAdminByUserKey(admin.getId())
-
                 ).doesNotThrowAnyException();
     }
